+++ src/app/search/components/index.ts
@@ -1,0 +1,1 @@
+export { default as SearchHistoryItem } from './SearchHistoryItem';
